+++ CHANGES.md
@@ -23,6 +23,7 @@ Activity subclasses maintain a field {:task_wrap_extensions} that can be used to
 * :pass_aggregate
 * Remove deprecated `:input`, `:output` and `:inject` option.
 * Remove deprecated `Out(with_outer_ctx: true) => ->(inner_ctx, outer_ctx, model:, **)` `outer_ctx` positional argument.
+* Remove deprecated `:override` option.
 
 # 1.2.7